SlideSpeak/pyproject.toml
[project]
name = "slidespeak"
version = "0.2.0"
description = "Convert PowerPoint presentations to video with voiceover - Desktop App with GUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-pptx>=0.6.23",
    "gtts>=2.5.1",
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "comtypes>=1.4.0",
    "pdf2image>=1.17.0",
    "flet>=0.21.0",
    "pydub>=0.25.1",
    "pytest>=8.0.0",
]

[project.optional-dependencies]
tts = ["coqui-tts>=0.20.0"]
build = ["flet[build]>=0.21.0", "pyinstaller>=6.0.0"]
dev = ["pytest>=8.0.0", "ruff>=0.1.0"]

[project.scripts]
slidespeak = "slidespeak.cli.main:main"
slidespeak-gui = "slidespeak.gui.main:start_gui"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["slidespeak"]

[tool.ruff]
line-length = 100
target-version = "py312"
